+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/cli
 
+## 0.5.88
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.40
+    - @stoker-platform/utils@0.5.52
+    - @stoker-platform/node-client@0.5.61
+
 ## 0.5.87
 
 ### Patch Changes