+++ packages/system-functions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/system-functions
 
+## 0.5.68
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.40
+    - @stoker-platform/utils@0.5.52
+    - @stoker-platform/node-client@0.5.61
+
 ## 0.5.67
 
 ### Patch Changes